book store db

+++ task_2.sql
@@ -30,7 +30,7 @@ CREATE TABLE Order_Details(
     orderdetailid INT PRIMARY KEY AUTO_INCREMENT,
     order_id INT,
     book_id INT,
-    quantity INT,
+    quantity DECIMAL(10,2),
     FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
